HW02-some functionality with controller

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryUserMealRepositoryImpl.java
@@ -1,11 +1,15 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Repository;
+import org.springframework.stereotype.Service;
 import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 import ru.javawebinar.topjava.util.UserMealsUtil;
 
-import java.util.Collection;
-import java.util.Map;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
 
@@ -13,12 +17,17 @@ import java.util.concurrent.atomic.AtomicInteger;
  * GKislin
  * 15.09.2015.
  */
+@Repository
 public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
+
     private Map<Integer, UserMeal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
 
     {
         UserMealsUtil.MEAL_LIST.forEach(this::save);
+    }
+
+    public InMemoryUserMealRepositoryImpl() {
     }
 
     @Override
@@ -31,18 +40,39 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public void delete(int id, int userId) {
+        if (repository.get(id).getUserId() == userId)
+            repository.remove(id);
     }
 
     @Override
-    public UserMeal get(int id) {
-        return repository.get(id);
+    public UserMeal get(int id, int userId) {
+
+        UserMeal result = repository.get(id);
+
+        if (result.getUserId() != userId)
+            result = null;
+
+        return result;
     }
 
     @Override
-    public Collection<UserMeal> getAll() {
-        return repository.values();
+    public List<UserMeal> getAll(int userId) {
+        List<UserMeal> result = new ArrayList<>();
+
+        repository.values().forEach(userMeal -> {
+            if (userMeal.getUserId() == userId)
+                result.add(userMeal);
+        });
+
+        if (!result.isEmpty())
+            Collections.sort(result, new Comparator<UserMeal>() {
+                @Override
+                public int compare(UserMeal o1, UserMeal o2) {
+                    return o1.getDateTime().compareTo(o2.getDateTime());
+                }
+            });
+
+        return result.isEmpty() ? null : result;
     }
 }
-
